In the load create screen pressing mod1 button(currently aura) takes … …you to a list of mods. clicking
on the mod on the list takes mod detail fragment.

Also I marked commented where the n/a is coming from. it's in MyWarframerecyclerViewAdapter.

Also added .finish() in signin activity.

Another thing I had to add a constructor in the mod class  and add constants inorder to extract information from the mod json.